@@ -1,6 +1,6 @@
 class Users::SessionsController < Devise::SessionsController
   skip_before_action :check_user_active_time
-  before_filter :configure_sign_in_params, only: [:create]
+  before_action :configure_sign_in_params, only: [:create]
   # GET /resource/sign_in
   # def new
   #   super
