@@ -18,7 +18,8 @@ Rails.application.configure do
 
     config.cache_store = :memory_store
     config.public_file_server.headers = {
-      'Cache-Control' => 'public, max-age=172800'
+      # 'Cache-Control' => 'public, max-age=172800'
+      'Cache-Control' => "public, max-age=#{2.days.seconds.to_i}"
     }
   else
     config.action_controller.perform_caching = false
@@ -57,4 +58,5 @@ Rails.application.configure do
   #
   #
   Paperclip.options[:command_path] = "/usr/bin/"
+  config.file_watcher = ActiveSupport::EventedFileUpdateChecker
 end
